requirements.txt
python-dotenv==1.2.2
playwright==1.52.0
python-telegram-bot==22.7
anthropic>=0.39.0
aiohttp==3.13.5
apscheduler==3.11.2
tenacity==9.1.4
